requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
boto==2.49.0
boto3==1.9.182
botocore==1.12.182
bz2file==0.98
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
click==7.1.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
docutils==0.14
gast==0.2.2
gensim==3.4.0
grpcio==1.22.0
h5py==2.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
jedi==0.13.3
jmespath==0.9.4
joblib==0.14.1
jupyter-client==5.2.4
jupyter-core==4.5.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.1.0
mock==3.0.5
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.6.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
pytz==2019.1
PyYAML==5.3.1
pyzmq==18.0.0
regex==2019.6.8
requests==2.22.0
s3transfer==0.2.1
sacremoses==0.0.38
scipy==1.3.1
six==1.12.0
smart-open==1.8.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.4.0
termcolor==1.1.0
torch==1.3.1
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
transformers==2.1.1
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.5
